# apps/backend/openapi.yaml
openapi: 3.1.0
info:
  title: OpenCloudTouch
  description: Open-Source replacement for discontinued streaming device cloud features
  version: dev-6b24c3b
paths:
...
  /bmx/radiobrowser/v1/playback/station/{uuid}:
    get:
      tags:
      - bmx
      summary: Bmx Radiobrowser Playback
      description: 'Resolve RadioBrowser station UUID to stream URL.


        The device calls this with a station UUID and expects a BMX-format

        JSON response with stream URLs for playback.'
      operationId: bmx_radiobrowser_playback_bmx_radiobrowser_v1_playback_station__uuid__get
      parameters:
      - name: uuid
        in: path
        required: true
        schema:
          type: string
          title: Uuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /bmx/radiobrowser/v1/now-playing/station/{uuid}:
    get:
      tags:
      - bmx
      summary: Bmx Radiobrowser Now Playing
      description: Now-playing stub for RadioBrowser stations.
      operationId: bmx_radiobrowser_now_playing_bmx_radiobrowser_v1_now_playing_station__uuid__get
      parameters:
      - name: uuid
        in: path
        required: true
        schema:
          type: string
          title: Uuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /bmx/radiobrowser/v1/reporting/station/{uuid}:
    post:
      tags:
      - bmx
      summary: Bmx Radiobrowser Reporting
      description: Reporting stub for RadioBrowser stations.
      operationId: bmx_radiobrowser_reporting_bmx_radiobrowser_v1_reporting_station__uuid__post
      parameters:
      - name: uuid
        in: path
        required: true
        schema:
          type: string
          title: Uuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...